Move Phase2L1GMT private headers in to plugins dir
