+++ L1Trigger/Phase2L1GMT/plugins/MuonROI.h
@@ -1,7 +1,7 @@
 #ifndef PHASE2GMT_MUONROI
 #define PHASE2GMT_MUONROI
 #include <iosfwd>
-#include "L1Trigger/Phase2L1GMT/interface/Constants.h"
+#include "Constants.h"
 #include "DataFormats/L1TMuonPhase2/interface/MuonStub.h"
 #include "DataFormats/L1TMuon/interface/RegionalMuonCand.h"
 #include "DataFormats/L1TMuon/interface/RegionalMuonCandFwd.h"
