+++ L1Trigger/Phase2L1GMT/plugins/Phase2L1TGMTSAMuonProducer.cc
@@ -37,7 +37,7 @@
 #include "DataFormats/L1Trigger/interface/L1MuonParticleFwd.h"
 #include "DataFormats/L1Trigger/interface/L1MuonParticle.h"
 
-#include "L1Trigger/Phase2L1GMT/interface/Constants.h"
+#include "Constants.h"
 #include "DataFormats/L1TMuonPhase2/interface/SAMuon.h"
 //
 // class declaration
