+++ L1Trigger/Phase2L1GMT/plugins/PreTrackMatchedMuon.h
@@ -2,7 +2,7 @@
 #define PHASE2GMT_PRETRACKMATCHEDMUON
 
 #include "FWCore/MessageLogger/interface/MessageLogger.h"
-#include "L1Trigger/Phase2L1GMT/interface/Constants.h"
+#include "Constants.h"
 #include "DataFormats/L1TMuon/interface/RegionalMuonCandFwd.h"
 #include "DataFormats/L1TMuonPhase2/interface/MuonStub.h"
 #include "DataFormats/Common/interface/Ptr.h"
