+++ L1Trigger/Phase2L1GMT/plugins/TopologicalAlgorithm.h
@@ -17,8 +17,8 @@
 #define PHASE2GMT_TOPOLOGICALALGORITHM
 
 #include "DataFormats/L1TMuonPhase2/interface/TrackerMuon.h"
-#include "L1Trigger/Phase2L1GMT/interface/ConvertedTTTrack.h"
-#include "L1Trigger/Phase2L1GMT/interface/Constants.h"
+#include "ConvertedTTTrack.h"
+//#include "Constants.h"
 
 #include <fstream>
