+++ L1Trigger/Phase2L1GMT/plugins/TrackConverter.h
@@ -1,7 +1,8 @@
 #ifndef PHASE2GMT_TRACKCONVERTER
 #define PHASE2GMT_TRACKCONVERTER
 
-#include "L1Trigger/Phase2L1GMT/interface/ConvertedTTTrack.h"
+#include "ConvertedTTTrack.h"
+#include "Constants.h"
 #include "DataFormats/L1TMuonPhase2/interface/TrackerMuon.h"
 #include "FWCore/ParameterSet/interface/ParameterSet.h"
